requirements.txt
flask-restful
numpy
pynamodb
s2sphere
pyyaml
waitress
